+++ src/main/kotlin/org/danilopianini/plagiarismdetector/input/configuration/RunConfiguration.kt
@@ -2,7 +2,7 @@ package org.danilopianini.plagiarismdetector.input.configuration
 
 import org.danilopianini.plagiarismdetector.core.TechniqueFacade
 import org.danilopianini.plagiarismdetector.core.detector.Match
-import org.danilopianini.plagiarismdetector.output.ReportsExporter
+import org.danilopianini.plagiarismdetector.output.exporter.ReportsExporter
 import org.danilopianini.plagiarismdetector.repository.Repository
 
 /**
